server-side caching implemented

+++ src/services/postgres/NotesService.js
@@ -8,9 +8,10 @@ const { mapDbToModel } = require('../../utils');
 const nanoid = customAlphabet('1234567890abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ', 16);
 
 class NotesService {
-  constructor(collaborationsService) {
+  constructor(collaborationsService, cacheService) {
     this._pool = new Pool();
     this._collaborationsService = collaborationsService;
+    this._cacheService = cacheService;
   }
 
   async addNote({
@@ -32,63 +33,91 @@ class NotesService {
       throw new InvariantError('Catatan gagal ditambahkan.');
     }
 
+    await this._cacheService.delete(`notes:${owner}`);
+
     return resultOk;
   }
 
   async getNotes(owner) {
-    const query = {
-      text: `SELECT notes.* FROM notes
-             LEFT JOIN collaborations ON collaborations.note_id = notes.id
-             WHERE notes.owner = $1 OR collaborations.user_id = $1
-             GROUP BY notes.id`,
-      values: [owner],
-    };
-    const result = await this._pool.query(query);
+    try {
+      const result = await this._cacheService.get(`notes:${owner}`);
+      return JSON.parse(result);
+    } catch (error) {
+      const query = {
+        text: `SELECT notes.* FROM notes
+              LEFT JOIN collaborations ON collaborations.note_id = notes.id
+              WHERE notes.owner = $1 OR collaborations.user_id = $1
+              GROUP BY notes.id`,
+        values: [owner],
+      };
+      const result = await this._pool.query(query);
+      const mappedResult = result.rows.map(mapDbToModel);
 
-    return result.rows.map(mapDbToModel);
+      await this._cacheService.set(`notes:${owner}`, JSON.stringify(mappedResult));
+
+      return mappedResult;
+    }
   }
 
   async getNoteById(id) {
-    const query = {
-      text: `SELECT notes.*, users.username
-             FROM notes
-             LEFT JOIN users ON users.id = notes.owner
-             WHERE notes.id = $1`,
-      values: [id],
-    };
-    const note = await this._pool.query(query);
-    if (!note.rows.length) {
-      throw new NotFoundError('Catatan tidak ditemukan.');
-    }
+    //--
+    try {
+      const result = await this._cacheService.get(`note:${id}`);
+      return JSON.parse(result);
+    } catch (error) {
+      const query = {
+        text: `SELECT notes.*, users.username
+              FROM notes
+              LEFT JOIN users ON users.id = notes.owner
+              WHERE notes.id = $1`,
+        values: [id],
+      };
+      const result = await this._pool.query(query);
 
-    return note.rows.map(mapDbToModel)[0];
+      if (!result.rowCount) {
+        throw new NotFoundError('Catatan tidak ditemukan.');
+      }
+
+      const mappedResult = result.rows.map(mapDbToModel)[0];
+      await this._cacheService.set(`note:${id}`, JSON.stringify(mappedResult)); //--
+
+      return mappedResult;
+    }
   }
 
   async editNoteById(id, { title, body, tags }) {
     const updatedAt = new Date().toISOString();
     const query = {
-      text: 'UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id',
+      text: 'UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id, owner',
       values: [title, body, tags, updatedAt, id],
     };
 
     const result = await this._pool.query(query);
 
-    if (!result.rows.length) {
+    if (!result.rowCount) {
       throw new NotFoundError('Gagal memperbarui catatan. Id tidak ditemukan.');
     }
+
+    const { id: noteId, owner } = result.rows[0];
+    await this._cacheService.delete(`notes:${owner}`);
+    await this._cacheService.delete(`note:${noteId}`); //--
   }
 
   async deleteNoteById(id) {
     const query = {
-      text: 'DELETE FROM notes WHERE id = $1 RETURNING id',
+      text: 'DELETE FROM notes WHERE id = $1 RETURNING id, owner',
       values: [id],
     };
 
     const result = await this._pool.query(query);
 
-    if (!result.rows.length) {
+    if (!result.rowCount) {
       throw new NotFoundError('Catatan gagal dihapus. Id tidak ditemukan.');
     }
+
+    const { id: noteId, owner } = result.rows[0];
+    await this._cacheService.delete(`notes:${owner}`);
+    await this._cacheService.delete(`note:${noteId}`); //--
   }
 
   async verifyNoteOwner(owner, id) {
